move SessionIdLogRecordAppender.kt into internal package
