+++ core/src/main/java/io/opentelemetry/android/internal/processors/SessionIdLogRecordAppender.kt
@@ -3,7 +3,7 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
-package io.opentelemetry.android
+package io.opentelemetry.android.internal.processors
 
 import io.opentelemetry.android.session.SessionProvider
 import io.opentelemetry.context.Context
